Zadanie_1/list_models.py
import os
from dotenv import load_dotenv
from google import genai

load_dotenv()
client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

for m in client.models.list():
    # v SDK býva zoznam podporovaných metód (napr. generateContent)
    methods = getattr(m, "supported_generation_methods", None)
    print(m.name, methods)